feat: Add Tari (RandomX-Tari) daemon mining support
Add support for mining on the Tari network to Tari base nodes with the xmrig_proxy enabled via its native daemon protocol.

Tari uses a 76-byte RandomX blob format that differs from standard Monero block templates, requiring dedicated parsing, hashing, and difficulty validation paths.

Algorithm & coin registration:

- Register RX_TARI algorithm (id 0x72151274, aliases "rx/tari", "randomxt")
- Add XTM coin info entry with 120s block time and 12-digit precision

RandomX configuration & hashing:

- Declare RandomX_ConfigurationTari using default base config (3 Argon iterations)
- New RxTari module: tari_randomx_calculate_hash() hashes the full 76-byte blob directly without Monero-format repacking

Block template parsing:

- Dedicated Tari path in BlockTemplate: parse 76-byte blob (3B header + 32B mining hash + 8B nonce LE + 33B pow data)
- Expose isTari(), tariMiningHash(), tariNonce(), tariPowData() accessors

Daemon client integration:

- Parse Tari-specific fields from getblocktemplate: min_nonce/max_nonce (u64 range), difficulty (u64 target)
- Retain last successful blocktemplate to avoid redundant requests on timeout
- Store daemon-assigned nonce range in Job struct with 8-byte nonce size for RX_TARI

Difficulty validation:

- achievesDifficulty() uses multi-precision multiplication (scalar * targetDiff <= U256::MAX)
- MSVC path uses \_umul128 intrinsics, GCC/Clang uses unsigned \__int128
- actualDiff() in JobResult uses full 256-bit formula for Tari

CPU worker hash loop:

- Early-exit Tari path before generic RandomX pipeline
- Direct tari_randomx_calculate_hash() call with daemon-assigned nonce boundary enforcement
- On job consumption, align per-worker starting nonces to daemon range using worker_id * N * kReserveCount offset

Job results & submission:

- Widened submit() overloads from uint32_t to uint64_t nonce parameter for 8-byte Tari nonces
- Log submitted nonce as 16-digit hex for Tari, 8-digit for all other algorithms

All changes are gated behind XMRIG_ALGO_RANDOMX / Algorithm::RX_TARI checks. Mining Monero or any existing algorithm is unaffected.

+++ src/net/JobResults.cpp
@@ -333,13 +333,13 @@ void xmrig::JobResults::stop()
 }
 
 
-void xmrig::JobResults::submit(const Job &job, uint32_t nonce, const uint8_t *result)
+void xmrig::JobResults::submit(const Job &job, uint64_t nonce, const uint8_t *result)
 {
     submit(JobResult(job, nonce, result));
 }
 
 
-void xmrig::JobResults::submit(const Job& job, uint32_t nonce, const uint8_t* result, const uint8_t* extra_data)
+void xmrig::JobResults::submit(const Job& job, uint64_t nonce, const uint8_t* result, const uint8_t* extra_data)
 {
     submit(JobResult(job, nonce, result, nullptr, nullptr, extra_data));
 }
